pipeline/requirements.txt
pyspark~=3.5.0
cassandra-driver~=3.28.0
python-dotenv~=1.0.0